Added Login/Register pages
