+++ budget/react_demo/src/pages/LoginPage.js
@@ -1,0 +1,9 @@
+import Login from "../Login/Login";
+
+const LoginPage = () => {
+  return (
+    <Login/>
+  );
+};
+
+export default LoginPage;
