+++ budget/react_demo/src/pages/NotFound.js
@@ -1,0 +1,9 @@
+const NotFound = () => {
+    return (
+      <div className="centered">
+        <p>Page not found!</p>
+      </div>
+    );
+  };
+  
+  export default NotFound;
