+++ budget/react_demo/src/pages/RegisterPage.js
@@ -1,0 +1,9 @@
+import Register from "../Login/Register";
+
+const RegisterPage = () => {
+  return (
+    <Register/>
+  );
+};
+
+export default RegisterPage;
